pyproject.toml
[tool.poetry]
name = "issue_dep_graph"
version = "0.2.0"
description = ""
authors = ["Srinivas Gorur-Shandilya <code@srinivas.gs>"]
readme = "README.md"
packages = [{include = "issue_dep_graph"}]

[tool.poetry.dependencies]
python = "^3.9"
pygithub = "^1.59.0"
beartype = "^0.15.0"
arguably = "^1.2.5"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.20.1"
debugpy = "1.6"


[tool.poetry.scripts]
idg = "issue_dep_graph.idg:start"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


